Unexport maybeAutoUpdate and split realRunInstaller by OS
Windows no longer pretends cmd.exe can run the installer; the stub
returns not implemented, and POSIX keeps sh -c.

Entire-Checkpoint: 01M1NYP4321A3WYS2ZAMAV6YBP
